requirements.txt
chainlit
langchain
langchain-community
langchain_core
langchain-openai
requests
pillow
python-dotenv